Code Restructered
