+++ app/src/main/java/org/acmvit/gitpositive/di/DatabaseModule.kt
@@ -1,4 +1,4 @@
-package org.acmvit.gitpositive
+package org.acmvit.gitpositive.di
 
 import android.content.Context
 import androidx.room.Room
