+++ app/src/main/java/org/acmvit/gitpositive/di/NetworkModule.kt
@@ -1,9 +1,10 @@
-package org.acmvit.gitpositive
+package org.acmvit.gitpositive.di
 
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import org.acmvit.gitpositive.remote.ApiInterface
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Singleton
