+++ app/src/main/java/org/acmvit/gitpositive/remote/model/RepositoryResponseItem.kt
@@ -1,4 +1,4 @@
-package org.acmvit.gitpositive.repositoryList.model
+package org.acmvit.gitpositive.remote.model
 
 data class RepositoryResponseItem(
     val allow_forking: Boolean,
