+++ app/src/main/java/org/acmvit/gitpositive/ui/repository/RepositoryViewModel.kt
@@ -1,4 +1,4 @@
-package org.acmvit.gitpositive.repositoryList.viewmodel
+package org.acmvit.gitpositive.ui.repository
 
 import android.util.Log
 import androidx.compose.runtime.MutableState
@@ -8,8 +8,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
-import org.acmvit.gitpositive.ApiInterface
-import org.acmvit.gitpositive.repositoryList.model.RepositoryResponseItem
+import org.acmvit.gitpositive.remote.ApiInterface
+import org.acmvit.gitpositive.remote.model.RepositoryResponseItem
 import javax.inject.Inject
 
 @HiltViewModel
