+++ app/src/main/res/layout/activity_about.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/bg_color"
-    tools:context=".repositoryList.AboutActivity">
+    tools:context=".ui.about.AboutActivity">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/share"
