+++ app/src/main/res/layout/activity_home_screen.xml
@@ -10,7 +10,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/bg_color"
-        tools:context=".HomeScreen">
+        tools:context=".ui.home.HomeScreen">
 
         <TextView
             android:id="@+id/Appname"
